+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/ArticleDtoMapper.java
@@ -1,10 +1,10 @@
-package nettee.article.driving.web;
+package nettee.blole.article.web;
 
 import nettee.blolet.article.domain.Article;
 import nettee.blolet.article.domain.ArticleLikes;
-import nettee.article.driving.web.dto.ArticleCommandDto.ArticleUpdateCommand;
-import nettee.article.driving.web.dto.ArticleCommandDto.ArticleCreateCommand;
-import nettee.article.driving.web.dto.ArticleLikesCommandDto.ArticleLikesCreateCommand;
+import nettee.blole.article.web.dto.ArticleCommandDto.ArticleUpdateCommand;
+import nettee.blole.article.web.dto.ArticleCommandDto.ArticleCreateCommand;
+import nettee.blole.article.web.dto.ArticleLikesCommandDto.ArticleLikesCreateCommand;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
