+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/ArticleQueryApi.java
@@ -1,10 +1,10 @@
-package nettee.article.driving.web;
+package nettee.blole.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import nettee.article.driving.web.dto.ArticleQueryDto.ArticleDetailViewResponse;
-import nettee.article.driving.web.dto.ArticleQueryDto.ArticleListViewsResponse;
+import nettee.blole.article.web.dto.ArticleQueryDto.ArticleDetailViewResponse;
+import nettee.blole.article.web.dto.ArticleQueryDto.ArticleListViewsResponse;
 import nettee.blolet.article.application.usecase.ArticleReadUseCase;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
