+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/dto/ArticleCommandDto.java
@@ -1,4 +1,4 @@
-package nettee.article.driving.web.dto;
+package nettee.blole.article.web.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
